requirements.txt
streamlit
requests
python-dotenv
opencv-python
bcrypt
facebook-business
streamlit-authenticator
